Move interactive read-eval-print out of runtime
This is not required for a server process like Gerrit Code Review.
Yanking it out of the runtime JAR slims the binary down by a few
hundred KiBs as the interactive debugger and associated code can
also be trimmed out.

Consulting a file in the interactive console now requires using

  consult('path/to/file.pl').

instead of the short-hand ["path/to/file.pl"] that many Prolog
implementations honor.

Change-Id: I5c8df131df349d5ca65d5684ed7cc1881247563b