Inline is{Integer,Double,Symbol,List,...} with instanceof
Most places that type-test a Term immediately downcast if the type
test is successful. The Java JIT is used to seeing "x instanceof Foo"
before "((Foo) x)" and can optimize the type-test and downcast logic
in machine code to only check the type once.

Take advantage of that by swapping out the isXxx() methods on Term
for proper instanceof checks in the Java code.

Flip the logic for $cut around to be more like human written code
where the common success path is the term is the correct type and
can be immediately downcast. JITs expect this flow as it happens
so often in Java source, and $cut happens often in the Prolog meta
interpreter.

Change-Id: I4ba34cab751b6060a3b1cc1adbf75ed0019b36fd