Add options to PrologMain to emulate usage in Gerrit
Disable features by default like IO and STATISTICS but
allow them to be enabled with flags on the command line.

Allow the caller to parse files outside of Prolog similar
to Gerrit by passing in `-f foo.pl` flags to consult from
a stream that was opened outside of Prolog.

Cap the database size at 256 by default, similar to Gerrit
capping the user database. Allow this to be raised using a
flag on the command line.

By default launch into the repl similar to SWI-Prolog.

Change-Id: Icaf1ef7ff3bd2f19a7ded4735e1fc945ef7bbbb6