Add split_java_sources.py script
This script works around two current limitations of Bazel[1] that come
up when trying to compile Prolog Cafe compiler output:

* Labels in BUILD files may not contain $, which is valid in a Java
  class name and thus in .java filenames.
* Bazel gets confused when a single .java source file contains multiple
  top-level classes.

This script takes as input the .java files produced by the Prolog Cafe
compiler, and produces a single .zip file as output, fixing the issues
above. This source jar can be passed directly as a src to a java_library
rule.

[1] https://github.com/bazelbuild/bazel/issues/374

Change-Id: I9cf2e9091a5416aa7512a51af751fa38e788a4cc