Update patch set 1
Patch Set 1:

This is GPL/EPL for now to match the surrounding project. If we decide to move it to another project (e.g. bazlets), we can relicense it to Apache as long as all contributors agree. (Currently Google is the only contributor and I wrote the whole thing; I'm guessing it's not going to accumulate too many external contributions :)

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>